Fix omission: Redirecting stderr to /dev/null
It turns out that in the test_% target of the Makefile, someone forgot to
exclude the redirection of stderr to /dev/null.

But wait! This modification was actually made during the exciting development
phase of the queue module. The tests are indeed passing.
For reasons unknown, I kept postponing pushing this fix.